A very poorly made test
Pushing this for the sole purpose of having it on Linux, this consistently logs the host address for that stream buffer KH3 reads bad data from.
Using this with the power of GDB should let me identify where the bad data comes from, but it's too late in the day for me to dig into this now.