Allow breaking out of inner loop when hitting end of file
In Batman: Return to Arkham titles, they try to exit AvPlayer while there are still video frames left. This causes a hang where the VideoDecoderThread is stuck on the m_video_buffers_cv while the DemuxerThread is trying to exit.

To fix this, add end of file checks to the inner loop in VideoDecoderThread, allowing it to exit early under these cases, instead of hanging.